Use the same spacing between members in decompilation that we do for Metadata-as-Source.
