+++ src/Features/CSharp/Portable/MetadataAsSource/CSharpMetadataAsSourceService.cs
@@ -10,8 +10,6 @@ using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.CodeGeneration;
 using Microsoft.CodeAnalysis.CSharp.DocumentationComments;
 using Microsoft.CodeAnalysis.CSharp.Simplification;
-using Microsoft.CodeAnalysis.CSharp.Utilities;
-using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.DocumentationComments;
 using Microsoft.CodeAnalysis.Formatting;
 using Microsoft.CodeAnalysis.Formatting.Rules;
@@ -22,10 +20,8 @@ using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CSharp.MetadataAsSource
 {
-    internal class CSharpMetadataAsSourceService : AbstractMetadataAsSourceService
+    internal partial class CSharpMetadataAsSourceService : AbstractMetadataAsSourceService
     {
-        private static readonly AbstractFormattingRule s_memberSeparationRule = new FormattingRule();
-
         public CSharpMetadataAsSourceService(HostLanguageServices languageServices)
             : base(languageServices.GetService<ICodeGenerationService>())
         {
@@ -55,9 +51,7 @@ namespace Microsoft.CodeAnalysis.CSharp.MetadataAsSource
         }
 
         protected override IEnumerable<AbstractFormattingRule> GetFormattingRules(Document document)
-        {
-            return s_memberSeparationRule.Concat(Formatter.GetDefaultFormattingRules(document));
-        }
+            => CSharpMetadataFormattingRule.Instance.Concat(Formatter.GetDefaultFormattingRules(document));
 
         protected override async Task<Document> ConvertDocCommentsToRegularCommentsAsync(Document document, IDocumentationCommentFormattingService docCommentFormattingService, CancellationToken cancellationToken)
         {
@@ -74,58 +68,5 @@ namespace Microsoft.CodeAnalysis.CSharp.MetadataAsSource
                 new CSharpEscapingReducer(),
                 new CSharpParenthesesReducer(),
                 new CSharpDefaultExpressionReducer());
-
-        private class FormattingRule : AbstractMetadataFormattingRule
-        {
-            protected override AdjustNewLinesOperation GetAdjustNewLinesOperationBetweenMembersAndUsings(SyntaxToken token1, SyntaxToken token2)
-            {
-                var previousToken = token1;
-                var currentToken = token2;
-
-                // We are not between members or usings if the last token wasn't the end of a statement or if the current token
-                // is the end of a scope.
-                if ((previousToken.Kind() != SyntaxKind.SemicolonToken && previousToken.Kind() != SyntaxKind.CloseBraceToken) ||
-                    currentToken.Kind() == SyntaxKind.CloseBraceToken)
-                {
-                    return null;
-                }
-
-                SyntaxNode previousMember = FormattingRangeHelper.GetEnclosingMember(previousToken);
-                SyntaxNode nextMember = FormattingRangeHelper.GetEnclosingMember(currentToken);
-
-                // Is the previous statement an using directive? If so, treat it like a member to add
-                // the right number of lines.
-                if (previousToken.Kind() == SyntaxKind.SemicolonToken && previousToken.Parent.Kind() == SyntaxKind.UsingDirective)
-                {
-                    previousMember = previousToken.Parent;
-                }
-
-                if (previousMember == null || nextMember == null || previousMember == nextMember)
-                {
-                    return null;
-                }
-
-                // If we have two members of the same kind, we won't insert a blank line 
-                if (previousMember.Kind() == nextMember.Kind())
-                {
-                    return FormattingOperations.CreateAdjustNewLinesOperation(1, AdjustNewLinesOption.ForceLines);
-                }
-
-                // Force a blank line between the two nodes by counting the number of lines of
-                // trivia and adding one to it.
-                var triviaList = token1.TrailingTrivia.Concat(token2.LeadingTrivia);
-                return FormattingOperations.CreateAdjustNewLinesOperation(GetNumberOfLines(triviaList) + 1, AdjustNewLinesOption.ForceLines);
-            }
-
-            public override void AddAnchorIndentationOperations(List<AnchorIndentationOperation> list, SyntaxNode node, AnalyzerConfigOptions options, in NextAnchorIndentationOperationAction nextOperation)
-            {
-                return;
-            }
-
-            protected override bool IsNewLine(char c)
-            {
-                return SyntaxFacts.IsNewLine(c);
-            }
-        }
     }
 }
